requirements.txt
beautifulsoup4==4.13.4
certifi==2025.1.31
charset-normalizer==3.4.1
faiss-cpu==1.10.0
filelock==3.18.0
fsspec==2025.3.2
huggingface-hub==0.30.2
idna==3.10
Jinja2==3.1.6
joblib==1.4.2
MarkupSafe==3.0.2
networkx==3.4.2
numpy==2.2.5
packaging==25.0
pandas==2.2.3
pillow==11.2.1
python-dateutil==2.9.0.post0
pytz==2025.2
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.3
safetensors==0.5.3
scikit-learn==1.6.1
scipy==1.15.2
sentence-transformers==4.1.0
setuptools==79.0.0
six==1.17.0
soupsieve==2.7
sympy==1.13.1
threadpoolctl==3.6.0
tokenizers==0.21.1
torch==2.6.0
torchvision==0.17.3
tqdm==4.67.1
transformers==4.51.3
typing_extensions==4.13.2
tzdata==2025.2
urllib3==2.4.0
openai
python-dotenv
litellm
gradio